Add files via upload
TASK 1
This code defines a function that calculates cost that takes the number of senior citizens as input and returns the total cost and cost per person. It checks if the number of seniors falls within the allowed range and calculates the cost based on the provided criteria. Finally, it prompts the user for input and displays the total cost and cost per person if the input is valid.
TASK 2
This code prompts the user to enter the names and payments for each senior citizen going on the outing and for any extra people. It then calculates the total amount collected and prints out the list of people on the outing along with the total amount collected.
TASK 3
This code determines whether the outing has made a profit or has broken even, we compare the total amount collected from Task 2 with the estimated cost from Task 1. If the total amount collected is greater than or equal to the estimated cost, the outing has broken even or made a profit; otherwise, it has incurred a loss.